Phase 8 item 8.4: Consolidate Settings and System pages
Move QueueStatusSection and SkillsSection from Settings.tsx to
System.tsx (which already had richer implementations from 8.2).
Port inline cron schedule editing to System.tsx SkillsTab.
Settings.tsx is now configuration-focused; System.tsx is the
operational hub. Fix unused extractMessageId in Email.tsx.
