+++ packages/web/src/pages/System.tsx
@@ -17,8 +17,12 @@ import {
   HardDrive,
   Database,
   Archive,
+  Pencil,
+  Check,
+  X,
 } from 'lucide-react';
 import { Button } from '@/components/ui/button';
+import { Input } from '@/components/ui/input';
 import { Badge } from '@/components/ui/badge';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Separator } from '@/components/ui/separator';
@@ -107,6 +111,19 @@ function describeCron(expr: string): string {
   }
 
   return expr;
+}
+
+/**
+ * Basic client-side cron expression validation.
+ * Accepts standard 5-field cron: minute hour day-of-month month day-of-week
+ */
+function isValidCron(expr: string): boolean {
+  const trimmed = expr.trim();
+  if (!trimmed) return false;
+  const fields = trimmed.split(/\s+/);
+  if (fields.length !== 5) return false;
+  const fieldPattern = /^(\*|[0-9]+(-[0-9]+)?(,[0-9]+(-[0-9]+)?)*)(\/([0-9]+))?$/;
+  return fields.every((f) => fieldPattern.test(f));
 }
 
 function statusColor(status: string): string {
@@ -269,15 +286,24 @@ function SkillsTab({
   loading,
   error,
   onTrigger,
+  onScheduleUpdate,
 }: {
   skills: Skill[];
   skillRuns: SkillLastRun[];
   loading: boolean;
   error: string | null;
   onTrigger: (name: string) => Promise<void>;
+  onScheduleUpdate: (name: string, schedule: string) => Promise<void>;
 }) {
   const [triggering, setTriggering] = useState<string | null>(null);
   const [triggerMsg, setTriggerMsg] = useState<Record<string, string>>({});
+
+  // Schedule editing state
+  const [editingSkill, setEditingSkill] = useState<string | null>(null);
+  const [editValue, setEditValue] = useState('');
+  const [editError, setEditError] = useState<string | null>(null);
+  const [saving, setSaving] = useState(false);
+  const [saveSuccess, setSaveSuccess] = useState<string | null>(null);
 
   // Build a map from skill_name -> last run info
   const runMap = new Map<string, SkillLastRun>();
@@ -295,6 +321,50 @@ function SkillsTab({
       setTriggerMsg((m) => ({ ...m, [name]: err instanceof Error ? err.message : 'Failed' }));
     } finally {
       setTriggering(null);
+    }
+  }
+
+  function handleEditClick(skillName: string, currentSchedule: string) {
+    setEditingSkill(skillName);
+    setEditValue(currentSchedule);
+    setEditError(null);
+    setSaveSuccess(null);
+  }
+
+  function handleCancel() {
+    setEditingSkill(null);
+    setEditValue('');
+    setEditError(null);
+  }
+
+  async function handleSave(skillName: string) {
+    const trimmed = editValue.trim();
+    if (!isValidCron(trimmed)) {
+      setEditError('Invalid cron expression. Expected 5 fields: minute hour day month weekday');
+      return;
+    }
+    setSaving(true);
+    setEditError(null);
+    try {
+      await onScheduleUpdate(skillName, trimmed);
+      setEditingSkill(null);
+      setEditValue('');
+      setSaveSuccess(skillName);
+      setTimeout(() => setSaveSuccess((prev) => prev === skillName ? null : prev), 3000);
+    } catch (err) {
+      setEditError(err instanceof Error ? err.message : 'Failed to update schedule');
+    } finally {
+      setSaving(false);
+    }
+  }
+
+  function handleKeyDown(e: React.KeyboardEvent, skillName: string) {
+    if (e.key === 'Enter') {
+      e.preventDefault();
+      handleSave(skillName);
+    } else if (e.key === 'Escape') {
+      e.preventDefault();
+      handleCancel();
     }
   }
 
@@ -326,6 +396,8 @@ function SkillsTab({
         const lastRunAt = lastRun?.last_run_at ?? skill.last_run_at ?? skill.last_run;
         const lastStatus = skill.last_run_status ?? (lastRunAt ? 'success' : undefined);
         const duration = lastRun?.duration_ms;
+        const isEditing = editingSkill === skill.name;
+        const justSaved = saveSuccess === skill.name;
 
         return (
           <div key={skill.name} className="px-4 py-3 flex items-center justify-between gap-4">
@@ -337,17 +409,69 @@ function SkillsTab({
                     ? <CheckCircle className="h-3.5 w-3.5 text-green-500 shrink-0" />
                     : <XCircle className="h-3.5 w-3.5 text-destructive shrink-0" />
                 )}
+                {justSaved && (
+                  <span className="text-xs text-green-600 dark:text-green-400">Schedule updated</span>
+                )}
               </div>
               <div className="text-xs text-muted-foreground mt-0.5 flex flex-wrap gap-x-3 gap-y-0.5">
-                {skill.schedule && (
-                  <span title={skill.schedule}>
-                    {describeCron(skill.schedule)}
+                {isEditing ? (
+                  <div className="mt-1 space-y-1.5" onClick={(e) => e.stopPropagation()}>
+                    <div className="flex items-center gap-1.5">
+                      <span className="text-muted-foreground shrink-0">Schedule:</span>
+                      <Input
+                        value={editValue}
+                        onChange={(e) => { setEditValue(e.target.value); setEditError(null); }}
+                        onKeyDown={(e) => handleKeyDown(e, skill.name)}
+                        className="w-48 h-6 px-2 font-mono text-xs"
+                        placeholder="0 20 * * 0"
+                        autoFocus
+                        disabled={saving}
+                      />
+                      <Button
+                        size="sm"
+                        variant="ghost"
+                        className="h-6 w-6 p-0 text-green-600 hover:text-green-700 hover:bg-green-50 dark:hover:bg-green-950"
+                        onClick={() => handleSave(skill.name)}
+                        disabled={saving}
+                        aria-label="Save schedule"
+                      >
+                        {saving ? <Loader2 className="h-3.5 w-3.5 animate-spin" /> : <Check className="h-3.5 w-3.5" />}
+                      </Button>
+                      <Button
+                        size="sm"
+                        variant="ghost"
+                        className="h-6 w-6 p-0 text-muted-foreground hover:text-foreground"
+                        onClick={handleCancel}
+                        disabled={saving}
+                        aria-label="Cancel editing"
+                      >
+                        <X className="h-3.5 w-3.5" />
+                      </Button>
+                    </div>
+                    {editValue.trim() && describeCron(editValue.trim()) !== editValue.trim() && (
+                      <p className="text-xs text-muted-foreground ml-[4.5rem]">{describeCron(editValue.trim())}</p>
+                    )}
+                    {editError && (
+                      <p className="text-xs text-destructive ml-[4.5rem]">{editError}</p>
+                    )}
+                  </div>
+                ) : (
+                  <span
+                    className="inline-flex items-center gap-1 group cursor-pointer hover:text-foreground transition-colors"
+                    onClick={() => handleEditClick(skill.name, skill.schedule ?? '')}
+                    role="button"
+                    tabIndex={0}
+                    onKeyDown={(e) => { if (e.key === 'Enter') handleEditClick(skill.name, skill.schedule ?? ''); }}
+                    title="Click to edit schedule"
+                  >
+                    {skill.schedule ? describeCron(skill.schedule) : 'No schedule'}
+                    <Pencil className="h-3 w-3 opacity-0 group-hover:opacity-60 transition-opacity" />
                   </span>
                 )}
-                {lastRunAt && (
+                {!isEditing && lastRunAt && (
                   <span>Last: {relativeTime(lastRunAt)}</span>
                 )}
-                {duration != null && (
+                {!isEditing && duration != null && (
                   <span>Duration: {formatDuration(duration)}</span>
                 )}
               </div>
@@ -1079,6 +1203,11 @@ export default function System() {
     await skillsApi.trigger(name);
   }
 
+  async function handleScheduleUpdate(name: string, schedule: string) {
+    await skillsApi.updateSchedule(name, schedule);
+    await loadSkills();
+  }
+
   // ── MCP load more ──
 
   function handleLoadMoreMcp() {
@@ -1196,6 +1325,7 @@ export default function System() {
           loading={skillsLoading}
           error={skillsError}
           onTrigger={handleTriggerSkill}
+          onScheduleUpdate={handleScheduleUpdate}
         />
       )}
